Move diff apply to workers
