+++ client/commands/state.go
@@ -7,6 +7,8 @@ import (
 
 	"github.com/ocuroot/ocuroot/client/release"
 	"github.com/ocuroot/ocuroot/client/state"
+	"github.com/ocuroot/ocuroot/client/tui"
+	"github.com/ocuroot/ocuroot/client/tui/tuiwork"
 	"github.com/ocuroot/ocuroot/client/work"
 	"github.com/ocuroot/ocuroot/refs"
 	"github.com/ocuroot/ocuroot/refs/refstore"
@@ -231,7 +233,17 @@ var StateApplyIntentCmd = &cobra.Command{
 
 		cmd.SilenceUsage = true
 
-		if err := state.ApplyIntent(ctx, tc.Ref, tc.State, tc.Intent); err != nil {
+		workTui := tui.StartWorkTui()
+		defer workTui.Cleanup()
+
+		tc.State = tuiwork.WatchForJobUpdates(ctx, tc.State, workTui)
+
+		worker := &work.Worker{
+			Tracker: tc,
+			Tui:     workTui,
+		}
+
+		if err := worker.ApplyIntent(ctx, tc.Ref); err != nil {
 			return fmt.Errorf("failed to apply intent: %w", err)
 		}
